[MIG] purchase_discount: Migration to 14.0
