+++ purchase_discount/models/purchase_order.py
@@ -3,6 +3,7 @@
 # Copyright 2015-2019 Tecnativa - Pedro M. Baeza
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
+
 from odoo import api, fields, models
 
 
@@ -10,8 +11,8 @@ class PurchaseOrder(models.Model):
     _inherit = "purchase.order"
 
     def _add_supplier_to_product(self):
-        """ Insert a mapping of products to PO lines to be picked up
-        in supplierinfo's create() """
+        """Insert a mapping of products to PO lines to be picked up
+        in supplierinfo's create()"""
         self.ensure_one()
         po_line_map = {
             line.product_id.product_tmpl_id.id: line for line in self.order_line
@@ -107,3 +108,30 @@ class PurchaseOrderLine(models.Model):
         vals = super(PurchaseOrderLine, self)._prepare_account_move_line(move)
         vals["discount"] = self.discount
         return vals
+
+    @api.model
+    def _prepare_purchase_order_line(
+        self, product_id, product_qty, product_uom, company_id, supplier, po
+    ):
+        """Apply the discount to the created purchase order"""
+        res = super()._prepare_purchase_order_line(
+            product_id, product_qty, product_uom, company_id, supplier, po
+        )
+        partner = supplier.name
+        uom_po_qty = product_uom._compute_quantity(product_qty, product_id.uom_po_id)
+        seller = product_id.with_company(company_id)._select_seller(
+            partner_id=partner,
+            quantity=uom_po_qty,
+            date=po.date_order and po.date_order.date(),
+            uom_id=product_id.uom_po_id,
+        )
+        res.update(self._prepare_purchase_order_line_from_seller(seller))
+        return res
+
+    @api.model
+    def _prepare_purchase_order_line_from_seller(self, seller):
+        """Overload this function to prepare other data from seller,
+        like in purchase_triple_discount module"""
+        if not seller:
+            return {}
+        return {"discount": seller.discount}
