+++ purchase_discount/static/description/index.html
@@ -367,7 +367,7 @@ ul.auto-toc {
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/purchase-workflow/tree/13.0/purchase_discount"><img alt="OCA/purchase-workflow" src="https://img.shields.io/badge/github-OCA%2Fpurchase--workflow-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/purchase-workflow-13-0/purchase-workflow-13-0-purchase_discount"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/142/13.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/purchase-workflow/tree/14.0/purchase_discount"><img alt="OCA/purchase-workflow" src="https://img.shields.io/badge/github-OCA%2Fpurchase--workflow-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/purchase-workflow-13-0/purchase-workflow-13-0-purchase_discount"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/142/14.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
 <p>This module allows to define a discount per line in the purchase orders. This
 discount can be also negative, interpreting it as an increment.</p>
 <p>It also modifies the purchase order report to include the discount field in it.</p>
@@ -378,16 +378,16 @@ it to purchase order lines:</p>
 discounted in price.</li>
 <li>You can set prices and discounts on the same screen.</li>
 </ul>
-<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/product_supplierinfo_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/product_supplierinfo_form.png" />
+<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/product_supplierinfo_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/product_supplierinfo_form.png" />
 <ul class="simple">
 <li>A new field default_supplierinfo_discount is added on res.partner model.
 This value will be used as the default one, on each supplierinfo of that
 supplier.</li>
 </ul>
-<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/res_partner_company_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/res_partner_company_form.png" />
+<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/res_partner_company_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/res_partner_company_form.png" />
 <p>Note: this setting is a new ‘company’ setting, unavailable for related
 partners, as accounting-related Settings.</p>
-<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/res_partner_individual_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/13.0/purchase_discount/static/description/res_partner_individual_form.png" />
+<img alt="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/res_partner_individual_form.png" src="https://raw.githubusercontent.com/OCA/purchase-workflow/14.0/purchase_discount/static/description/res_partner_individual_form.png" />
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -421,7 +421,7 @@ this field. So this module breaks all the other modules that use the <em>price_u
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/purchase-workflow/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/purchase-workflow/issues/new?body=module:%20purchase_discount%0Aversion:%2013.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/purchase-workflow/issues/new?body=module:%20purchase_discount%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -453,6 +453,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <li>Sudhir Arya &lt;<a class="reference external" href="mailto:sudhir&#64;erpharbor.com">sudhir&#64;erpharbor.com</a>&gt;</li>
 <li>Lorenzo Battistini &lt;<a class="reference external" href="https://github.com/eLBati">https://github.com/eLBati</a>&gt;</li>
 <li>Luisa Miguéns &lt;<a class="reference external" href="mailto:luisa.miguens&#64;solvos.es">luisa.miguens&#64;solvos.es</a>&gt;</li>
+<li>Vishnu Vanneri &lt;<a class="reference external" href="mailto:vanneri.odoodev&#64;gmail.com">vanneri.odoodev&#64;gmail.com</a>&gt;</li>
 </ul>
 </div>
 <div class="section" id="maintainers">
@@ -462,7 +463,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <p>OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/purchase-workflow/tree/13.0/purchase_discount">OCA/purchase-workflow</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/purchase-workflow/tree/14.0/purchase_discount">OCA/purchase-workflow</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
